refactor:StickyTriSectionHeader 분리
