+++ frontend/src/layouts/components/StickyTriSectionHeader/StickyTriSectionHeader.style.ts
